App/requirements.txt
python-dotenv==1.2.1
psycopg2-binary==2.9.11